requirements.txt
fuzzywuzzy==0.18.0
jieba==0.42.1
openai==0.28.0
pandas==2.3.1
python-dotenv==1.1.1
python-telegram-bot==22.3
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.1
selenium==4.34.2
streamlit==1.47.1
webdriver_manager==4.0.2
